Move JSON ingestion into elasticgraph-json_ingestion via an indexer extension
The JSON ingestion adapter and JSON-schema-based record preparer factory
move from elasticgraph-indexer to elasticgraph-json_ingestion, and the
JSON ingestion schema definition extension registers an indexer extension
in runtime metadata that makes the adapter available at indexing time.
Any schema defined with JSON ingestion support automatically gets JSON
event ingestion--no settings needed--and elasticgraph-indexer no longer
contains (or refers to) any JSON-specific ingestion logic.

The be_a_valid_elastic_graph_event matcher moves to
elastic_graph/json_ingestion/spec_support/event_matcher, and suites that
build indexers from the shared test schema artifacts gain a development
dependency on elasticgraph-json_ingestion (whose indexer extension those
artifacts now reference).
