+++ elasticgraph-indexer/sig/elastic_graph/indexer.rbs
@@ -19,17 +19,13 @@ module ElasticGraph
     @datastore_router: Indexer::_DatastoreRouter?
     def datastore_router: () -> Indexer::_DatastoreRouter
 
-    @record_preparer_factory: RecordPreparer::Factory?
-    def record_preparer_factory: () -> RecordPreparer::Factory
+    def ingestion_adapters: () -> ::Array[IngestionAdapter::_Adapter]
 
     @processor: Processor?
     def processor: () -> Processor
 
     @operation_factory: Operation::Factory?
     def operation_factory: () -> Operation::Factory
-
-    @ingestion_adapters: ::Array[IngestionAdapter::_Adapter]?
-    def ingestion_adapters: () -> ::Array[IngestionAdapter::_Adapter]
 
     @monotonic_clock: Support::MonotonicClock?
     def monotonic_clock: () -> Support::MonotonicClock
