--- elasticgraph-indexer/sig/elastic_graph/indexer/ingestion_adapter/json_events.rbs
@@ -1,31 +1,0 @@
-module ElasticGraph
-  class Indexer
-    module IngestionAdapter
-      class JSONEvents
-        def initialize: (
-          schema_artifacts: schemaArtifacts,
-          logger: ::Logger,
-          ?configure_record_validator: (^(Support::JSONSchema::ValidatorFactory) -> Support::JSONSchema::ValidatorFactory)?
-        ) -> void
-
-        @schema_artifacts: schemaArtifacts
-        @logger: ::Logger
-        @configure_record_validator: (^(Support::JSONSchema::ValidatorFactory) -> Support::JSONSchema::ValidatorFactory)?
-        @record_preparer_factory: RecordPreparer::Factory
-
-        def handles_event?: (event) -> bool
-        def validate_event: (event) -> ValidationResult
-
-        private
-
-        def select_json_schema_version: (event) { (ValidationResult) -> bot } -> (::Integer | bot)
-        def validator: (::String, ::Integer) -> Support::JSONSchema::Validator
-
-        @validator_factories_by_version: ::Hash[::Integer, Support::JSONSchema::ValidatorFactory]?
-        def validator_factories_by_version: () -> ::Hash[::Integer, Support::JSONSchema::ValidatorFactory]
-
-        def raise: (*untyped) -> bot
-      end
-    end
-  end
-end
