+++ elasticgraph-indexer/sig/elastic_graph/indexer/record_preparer.rbs
@@ -8,20 +8,6 @@ module ElasticGraph
       type egMetaByFieldHash = ::Hash[::String, {"type" => ::String, "nameInIndex" => ::String}]
       type egMetaByFieldByTypeHash = ::Hash[::String, egMetaByFieldHash]
 
-      class Factory
-        def initialize: (schemaArtifacts) -> void
-
-        def for_json_schema_version: (::Integer) -> RecordPreparer
-        def for_latest_json_schema_version: () -> RecordPreparer
-
-        private
-
-        @schema_artifacts: schemaArtifacts
-        @preparers_by_json_schema_version: ::Hash[::Integer, RecordPreparer]
-
-        def build_type_metas_from: (::Hash[::String, untyped]) -> ::Array[TypeMetadata]
-      end
-
       include _RecordPreparer
 
       module Identity
