+++ elasticgraph-json_ingestion/elasticgraph-json_ingestion.gemspec
@@ -36,6 +36,15 @@ Gem::Specification.new do |spec|
 
   spec.add_dependency "elasticgraph-support", ElasticGraph::VERSION
 
+  # The test suite builds indexers (with real datastore client classes) to exercise the ingestion adapter.
+  spec.add_development_dependency "elasticgraph-elasticsearch", ElasticGraph::VERSION
+  # This gem's ingestion adapter code references `elasticgraph-indexer`, but the indexer loads it
+  # through the indexer extension registered in runtime metadata, after `elasticgraph-indexer` is
+  # already available. Keeping this as a development dependency avoids a runtime dependency cycle.
+  spec.add_development_dependency "elasticgraph-indexer", ElasticGraph::VERSION
+  # The test suite builds indexers (with real datastore client classes) to exercise the ingestion adapter.
+  spec.add_development_dependency "elasticgraph-opensearch", ElasticGraph::VERSION
+
   # This gem's schema-definition extension code references `elasticgraph-schema_definition`, but
   # applications load it through schema-definition tasks after `elasticgraph-schema_definition` is already
   # available. Keeping this as a development dependency avoids a runtime dependency cycle.
