+++ elasticgraph-json_ingestion/sig/elastic_graph/json_ingestion/record_preparer_factory.rbs
@@ -1,0 +1,17 @@
+module ElasticGraph
+  module JSONIngestion
+    class RecordPreparerFactory
+      def initialize: (schemaArtifacts) -> void
+
+      def for_json_schema_version: (::Integer) -> Indexer::RecordPreparer
+      def for_latest_json_schema_version: () -> Indexer::RecordPreparer
+
+      private
+
+      @schema_artifacts: schemaArtifacts
+      @preparers_by_json_schema_version: ::Hash[::Integer, Indexer::RecordPreparer]
+
+      def build_type_metas_from: (::Hash[::String, untyped]) -> ::Array[Indexer::RecordPreparer::TypeMetadata]
+    end
+  end
+end
