+++ elasticgraph-local/lib/elastic_graph/local/spec_support/common_project_specs.rb
@@ -10,7 +10,7 @@ require "elastic_graph/admin"
 require "elastic_graph/graphql"
 require "elastic_graph/indexer"
 require "elastic_graph/indexer/test_support/converters"
-require "elastic_graph/indexer/spec_support/event_matcher"
+require "elastic_graph/json_ingestion/spec_support/event_matcher"
 require "elastic_graph/support/json_schema/validator"
 require "factory_bot"
 require "json_schemer"
