+++ elasticgraph-warehouse_lambda/elasticgraph-warehouse_lambda.gemspec
@@ -48,5 +48,8 @@ Gem::Specification.new do |spec|
 
   spec.add_development_dependency "aws_lambda_ric", "~> 3.2"
   spec.add_development_dependency "elasticgraph-elasticsearch", ElasticGraph::VERSION
+  # The test suite builds indexers from the shared test schema artifacts, whose runtime metadata
+  # registers the JSON ingestion indexer extension provided by `elasticgraph-json_ingestion`.
+  spec.add_development_dependency "elasticgraph-json_ingestion", ElasticGraph::VERSION
   spec.add_development_dependency "elasticgraph-opensearch", ElasticGraph::VERSION
 end
