+++ spec_support/lib/elastic_graph/spec_support/builds_indexer_operation.rb
@@ -32,7 +32,7 @@ module ElasticGraph
 
         Indexer::Operation::Update.new(
           event: event,
-          prepared_record: idxr.record_preparer_factory.for_latest_json_schema_version.prepare_for_index(
+          prepared_record: latest_json_record_preparer_for(idxr).prepare_for_index(
             event.fetch("type"),
             event.fetch("record"),
             idxr.schema_artifacts.index_mappings_by_index_def_name.fetch(index_def.name).fetch("properties")
